refactor: per-service exceptions
Moves exception types for each service into their own namespace, and
introduces an exception marker-interface specific to each service, to
allow catching exceptions by service (as well as individually).
