+++ src/Service/CoverallsService/CoverallsServiceExceptionInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Laminas\Transfer\Service\CoverallsService;
+
+use Laminas\Transfer\Exception\ExceptionInterface;
+
+interface CoverallsServiceExceptionInterface extends ExceptionInterface
+{
+}
