+++ src/Service/PackagistService/PackagistServiceExceptionInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Laminas\Transfer\Service\PackagistService;
+
+use Laminas\Transfer\Exception\ExceptionInterface;
+
+interface PackagistServiceExceptionInterface extends ExceptionInterface
+{
+}
